pyproject.toml
[tool.poetry]
name = "skybank-project-ovsianik"
version = "0.1.0"
description = ""
authors = ["Valentina Ovsianik <forovp@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
chardet = "^5.2.0"
charset-normalizer = "^3.3.2"


[tool.poetry.group.lint.dependencies]
mypy = "^1.10.1"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = """
(
  /(
      \\.eggs
    | \\.git
    | \\.hg
    | \\.mypy_cache
    | \\.tox
    | \\.venv
    | dist
  )/
  | foo.py # Также отдельно исключить файл с именем foo.py
)
"""

[tool.isort]
# Максимальная длина строки
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

